clean :common:features:tasks module
