+++ android/screens/src/main/kotlin/work/racka/reluct/android/screens/tasks/done/CompletedTasksScreen.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.launch
 import work.racka.common.mvvm.koin.compose.getCommonViewModel
 import work.racka.reluct.android.screens.R
-import work.racka.reluct.common.features.tasks.completed_tasks.CompletedTasksViewModel
+import work.racka.reluct.common.features.tasks.completedTasks.CompletedTasksViewModel
 import work.racka.reluct.common.model.states.tasks.TasksEvents
 import work.racka.reluct.compose.common.components.util.BarsVisibility
