+++ android/screens/src/main/kotlin/work/racka/reluct/android/screens/tasks/pending/PendingTasksScreen.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.launch
 import work.racka.common.mvvm.koin.compose.getCommonViewModel
 import work.racka.reluct.android.screens.R
-import work.racka.reluct.common.features.tasks.pending_tasks.PendingTasksViewModel
+import work.racka.reluct.common.features.tasks.pendingTasks.PendingTasksViewModel
 import work.racka.reluct.common.model.states.tasks.TasksEvents
 import work.racka.reluct.compose.common.components.util.BarsVisibility
