+++ common/features/tasks/src/commonTest/kotlin/work/racka/reluct/common/features.tasks/completedTasks/CompletedTasksTest.kt
@@ -1,4 +1,4 @@
-package work.racka.reluct.common.features.tasks.pending_tasks
+package work.racka.reluct.common.features.tasks.completedTasks
 
 import app.cash.turbine.test
 import io.mockk.MockKAnnotations
@@ -23,14 +23,14 @@ import work.racka.reluct.common.domain.mappers.tasks.asTask
 import work.racka.reluct.common.domain.usecases.tasks.GetTasksUseCase
 import work.racka.reluct.common.domain.usecases.tasks.ModifyTaskUseCase
 import work.racka.reluct.common.features.tasks.util.TestData
-import work.racka.reluct.common.model.states.tasks.PendingTasksState
+import work.racka.reluct.common.model.states.tasks.CompletedTasksState
 import work.racka.reluct.common.model.states.tasks.TasksEvents
 import kotlin.test.*
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class PendingTasksTest : KoinTest {
+class CompletedTasksTest : KoinTest {
 
-    private val pendingTasks: PendingTasks by inject()
+    private val completedTasks: CompletedTasks by inject()
 
     @RelaxedMockK
     private lateinit var getTasksUseCase: GetTasksUseCase
@@ -45,8 +45,8 @@ class PendingTasksTest : KoinTest {
         startKoin {
             modules(
                 module {
-                    factory<PendingTasks> {
-                        PendingTasksImpl(
+                    factory<CompletedTasks> {
+                        CompletedTasksImpl(
                             getTasksUseCase = getTasksUseCase,
                             modifyTasksUsesCase = modifyTasksUsesCase,
                             scope = CoroutineScope(StandardTestDispatcher())
@@ -64,32 +64,27 @@ class PendingTasksTest : KoinTest {
     }
 
     @Test
-    fun getPendingTasks_OnClassInit_ShouldUpdateUIStateWithCompletedTasks() =
+    fun getCompletedTasks_OnClassInit_ShouldUpdateUIStateWithCompletedTasks() =
         runTest {
             val factor = 1L
-            val taskList = TestData.taskDbObjects.map { it.asTask() }
-            val overdueList = taskList.filter { it.overdue }
-            val grouped = taskList
-                .filterNot { it.overdue }
-                .groupBy { it.dueDate }
-            val expectedState = PendingTasksState.Data(
-                tasks = grouped,
-                overdueTasks = overdueList
+            val tasks = TestData.taskDbObjects.map { it.asTask() }
+            val expectedState = CompletedTasksState.Data(
+                tasks = tasks.groupBy { it.dueDate }
             )
 
-            coEvery { getTasksUseCase.getPendingTasks(factor) } returns flowOf(taskList)
+            coEvery { getTasksUseCase.getCompletedTasks(factor) } returns flowOf(tasks)
 
-            val result = pendingTasks.uiState
+            val result = completedTasks.uiState
             launch {
                 result.test {
                     val initial = awaitItem()
                     val actual = awaitItem()
                     println(actual)
 
-                    assertTrue(initial is PendingTasksState.Loading)
-                    assertTrue(actual is PendingTasksState.Data)
+                    assertTrue(initial is CompletedTasksState.Loading)
+                    assertTrue(actual is CompletedTasksState.Data)
                     assertEquals(expectedState, actual)
-                    coVerify { getTasksUseCase.getPendingTasks(factor) }
+                    coVerify { getTasksUseCase.getCompletedTasks(factor) }
                     awaitComplete()
                 }
             }
@@ -106,10 +101,10 @@ class PendingTasksTest : KoinTest {
             val expectedEvent = TasksEvents.ShowMessageDone(isDone, task.title)
             coEvery { modifyTasksUsesCase.toggleTaskDone(task, isDone) } returns Unit
 
-            val result = pendingTasks.events
+            val result = completedTasks.events
             launch {
                 result.test {
-                    pendingTasks.toggleDone(task, isDone)
+                    completedTasks.toggleDone(task, isDone)
                     val actual = expectMostRecentItem()
                     println(actual)
 
@@ -126,10 +121,10 @@ class PendingTasksTest : KoinTest {
             val expectedEvent = TasksEvents.Navigation.NavigateToTaskDetails(taskId)
 
 
-            val result = pendingTasks.events
+            val result = completedTasks.events
             launch {
                 result.test {
-                    pendingTasks.navigateToTaskDetails(taskId)
+                    completedTasks.navigateToTaskDetails(taskId)
                     val actual = expectMostRecentItem()
                     println(actual)
